muita linha de codigo. no entanto os comandos ja funcionam com pipes, seja la quantos forem. Net step o resto dos caraceres especiais. tive q mudar algumas cenas q estava, a dar leaks no parser e adicionei umas verificacoes que ainda tem que ser aprofundadas

+++ libft/libft.h
@@ -57,7 +57,7 @@ char			*ft_strdup(const char	*s1);
 char			*ft_strjoin(char const *s1, char const *s2, char const *s3);
 size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
 size_t			ft_strlcpy(char *dest, const char *src, size_t size);
-size_t			ft_strlen(const char	*str);
+int				ft_strlen(const char	*str);
 char			*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 int				ft_strncmp(const char *s1, const char *s2, size_t n);
 int				ft_strcmp(const char *s1, const char *s2);
